Extracted SudokuTransformator class
